updated style on profile
